+++ src/actions/identityActions.js
@@ -198,35 +198,6 @@ export function setIdentityENS({ username }) {
 	}
 }
 
-export function identityWithdraw({ amountToWithdraw, withdrawTo }) {
-	return async function(dispatch, getState) {
-		try {
-			const account = selectAccount(getState())
-			const result = await withdrawFromIdentity({
-				account,
-				amountToWithdraw,
-				withdrawTo,
-			})
-			addToast({
-				type: 'accept',
-				label: translate('IDENTITY_WITHDRAW_NOTIFICATION', {
-					args: [result],
-				}),
-				timeout: 20000,
-			})(dispatch)
-		} catch (err) {
-			console.error('ERR_IDENTITY_WITHDRAW_NOTIFICATION', err)
-			addToast({
-				type: 'cancel',
-				label: translate('ERR_IDENTITY_WITHDRAW_NOTIFICATION', {
-					args: [getErrorMsg(err)],
-				}),
-				timeout: 20000,
-			})(dispatch)
-		}
-	}
-}
-
 export function identityWithdrawAny({
 	amountToWithdraw,
 	withdrawTo,
