+++ src/components/dashboard/forms/web3/transactions.js
@@ -4,11 +4,8 @@ import WithdrawAnyTokenFromIdentityPage from './WithdrawAnyTokenFromIdentity'
 import SeAddressPrivilege from './SeAddressPrivilege'
 import SetAccountENSPage from './SetAccountENSPage'
 import TransactionPreview from './TransactionPreview'
-import Button from '@material-ui/core/Button'
-import TransactionHoc from './TransactionHoc'
 import FormSteps from 'components/common/stepper/FormSteps'
 import WithDialog from 'components/common/dialog/WithDialog'
-import HourglassEmptyIcon from '@material-ui/icons/HourglassEmpty'
 import {
 	addIdentityENS,
 	withdrawOtherTokensFromIdentity,
@@ -31,40 +28,7 @@ const cancelFunction = stepsId => {
 	execute(resetNewTransaction({ tx: stepsId }))
 }
 
-const SaveBtn = ({
-	save,
-	saveBtnLabel,
-	saveBtnIcon,
-	t,
-	transaction,
-	waitingForWalletAction,
-	spinner,
-	...other
-}) => {
-	return (
-		<Button
-			color='primary'
-			onClick={save}
-			disabled={
-				(!!transaction.errors && !!transaction.errors.length) ||
-				!!transaction.waitingForWalletAction ||
-				!!spinner
-			}
-		>
-			{transaction.waitingForWalletAction ? (
-				<HourglassEmptyIcon />
-			) : (
-				saveBtnIcon || ''
-			)}
-			{t(saveBtnLabel || 'DO_IT')}
-		</Button>
-	)
-}
-
-const SaveBtnWithTransaction = TransactionHoc(SaveBtn)
-
 const txCommon = {
-	SaveBtn: SaveBtnWithTransaction,
 	cancelFunction,
 	darkerBackground: true,
 }
@@ -92,19 +56,21 @@ export const WithdrawTokenFromIdentity = props => (
 						})
 					),
 			},
+			{
+				title: 'PREVIEW_AND_MAKE_TR',
+				component: TransactionPreview,
+				completeFn: ({ stepsId, validateId, onValid, onInvalid }) =>
+					execute(
+						completeTx({
+							stepsId,
+							validateId,
+							competeAction: identityWithdraw,
+							onValid,
+							onInvalid,
+						})
+					),
+			},
 		]}
-		stepsPreviewPage={{
-			title: 'PREVIEW_AND_MAKE_TR',
-			page: TransactionPreview,
-		}}
-		saveFn={({ transaction: { amountToWithdraw, withdrawTo } } = {}) => {
-			return execute(
-				identityWithdraw({
-					amountToWithdraw,
-					withdrawTo,
-				})
-			)
-		}}
 	/>
 )
